added more tests for courses, added courseAssement tests

+++ App/controllers/course.py
@@ -28,7 +28,10 @@ def edit_course(review, staff, is_positive, comment):
         return review
     return None    
 
-def delete_Course(course):
-    db.session.delete(course)
-    db.session.commit()
-    return True     
+def delete_Course(courseCode):
+    course = Course.query.get(courseCode)
+    if course:
+        db.session.delete(course)
+        db.session.commit()
+        return True
+    return False
